test: add new test that uses OP_PUSHDATA1 and move script data tests to new file
